# Optimize LZMA range decoder